move loader and get_dir into helpers module
