+++ src/main.rs
@@ -1,9 +1,8 @@
-use std::env;
-use std::path::PathBuf;
 use structopt::StructOpt;
 
 mod clean;
 mod error;
+mod helpers;
 mod new;
 mod open;
 mod watch;
@@ -48,12 +47,6 @@ struct EditorOpts {
     pub args: Vec<String>,
 }
 
-fn get_dir() -> PathBuf {
-    env::var_os("CARGO_PLAYGROUND_DIR")
-        .map(PathBuf::from)
-        .unwrap_or_else(|| env::temp_dir().join("cargo-playground"))
-}
-
 fn main() {
     match run() {
         Ok(()) => {}
@@ -78,7 +71,7 @@ fn run() -> error::Result<()> {
         PlaygroundOpts::Open(opts) => open::open(opts),
         PlaygroundOpts::Clean(opts) => clean::clean(opts),
         PlaygroundOpts::Ls => {
-            let path = get_dir();
+            let path = helpers::get_dir();
             if !path.exists() {
                 return Ok(());
             }
